ENGR00175881 Mass blocks flashes when switching from Panorama to other mode.
Root cause, when switching from Panorama to other mode, the Panorama activity
was finished. But the preview surface was not desdroy at the same time. There
still was more frame queued. In this condition, the issue happened.
The way to fix it, set the preview surface to null and not display before
finishing the activity.

Signed-off-by: YangYonggang <b31664@freescale.com>